requirements.txt
appnope==0.1.0
decorator==4.0.9
gnureadline==6.3.3
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
simplegeneric==0.8.1
suds-py3==1.2.0.0
traitlets==4.1.0
wheel==0.24.0
